// src/lib/openapi/adminapi.schema.ts
export interface components {
    schemas: {
// ...
        CreateLegalEntityCorporate: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'corporate';
            kyb_status: components['schemas']['KybStatus'];
            onboarding_mode: components['schemas']['OnboardingMode'];
            /** @example Example Corp */
            company_name: string;
            /** @example [
             *       "Example Corp"
             *     ] */
            alias_names: string[];
            /** @example [
             *       "Example Corp"
             *     ] */
            trade_names: string[];
            /** @example [
             *       "Example Corp"
             *     ] */
            former_registered_names: string[];
            incorporation_date: components['schemas']['Date'];
            /** @example 12345678 */
            incorporation_number: string;
            incorporation_country: components['schemas']['CountryCode2'];
            operation_countries: components['schemas']['CountryCode2'][];
            /** @example [
             *       "https://example.com",
             *       "https://example.org"
             *     ] */
            websites: string[];
            /** @example 1 */
            industry_id: number;
            /** @example [
             *       "1234",
             *       "5678"
             *     ] */
            mcc_codes: string[];
            /** @example [
             *       "software development",
             *       "consulting"
             *     ] */
            business_activities: string[];
            /** @example [
             *       "investors",
             *       "revenue"
             *     ] */
            sources_of_fund: string[];
            addresses: components['schemas']['CreateEntityAddress'][];
        };
        /** CreateLegalEntityIndividual */
        CreateLegalEntityIndividual: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'individual';
            kyb_status: components['schemas']['KybStatus'];
            onboarding_mode: components['schemas']['OnboardingMode'];
            /** @example John */
            first_name: string;
            /** @example Doe */
            last_name: string;
            /** @example [
             *       "John D"
             *     ] */
            alias_names: string[];
            gender: components['schemas']['Gender'];
            nationalities: components['schemas']['CountryCode2'][];
            birth_country: components['schemas']['CountryCode2'];
            birth_date: components['schemas']['Date'];
            residence_country: components['schemas']['CountryCode2'];
            /** @example International Passport */
            identity_document_type: string;
            /** @example 123456 */
            identity_document_number: string;
            identity_document_issue_date: components['schemas']['Date'];
            identity_document_expiry_date: components['schemas']['Date'];
            /** @example 1 */
            occupation_id: number;
            /** @example [
             *       "investors",
             *       "revenue"
             *     ] */
            sources_of_fund: string[];
            addresses: components['schemas']['CreateEntityAddress'][];
        };
        CreateLegalEntity:
            | components['schemas']['CreateLegalEntityIndividual']
            | components['schemas']['CreateLegalEntityCorporate'];
        /** UpdateLegalEntityCorporate */
        UpdateLegalEntityCorporate: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'corporate';
            kyb_status: components['schemas']['KybStatus'];
            onboarding_mode: components['schemas']['OnboardingMode'];
            /** @example Example Corp */
            company_name: string;
            /** @example [
             *       "Example Corp"
             *     ] */
            alias_names: string[];
            /** @example [
             *       "Example Corp"
             *     ] */
            trade_names: string[];
            /** @example [
             *       "Example Corp"
             *     ] */
            former_registered_names: string[];
            incorporation_date: components['schemas']['Date'];
            /** @example 12345678 */
            incorporation_number: string;
            incorporation_country: components['schemas']['CountryCode2'];
            operation_countries: components['schemas']['CountryCode2'][];
            /** @example [
             *       "https://example.com",
             *       "https://example.org"
             *     ] */
            websites: string[];
            /** @example Technology */
            industry_id: number;
            /** @example [
             *       "1234",
             *       "5678"
             *     ] */
            mcc_codes: string[];
            /** @example [
             *       "Software Development",
             *       "Consulting"
             *     ] */
            business_activities: string[];
            /** @example [
             *       "Investors",
             *       "Revenue"
             *     ] */
            sources_of_fund: string[];
            addresses: components['schemas']['UpdateEntityAddress'][];
        };
        /** UpdateLegalEntityIndividual */
        UpdateLegalEntityIndividual: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            entity_type: 'individual';
            kyb_status: components['schemas']['KybStatus'];
            onboarding_mode: components['schemas']['OnboardingMode'];
            /** @example John */
            first_name: string;
            /** @example Doe */
            last_name: string;
            /** @example [
             *       "John D"
             *     ] */
            alias_names: string[];
            gender: components['schemas']['Gender'];
            nationalities: components['schemas']['CountryCode2'][];
            birth_country: components['schemas']['CountryCode2'];
            birth_date: components['schemas']['Date'];
            residence_country: components['schemas']['CountryCode2'];
            /** @example International Passport */
            identity_document_type: string;
            /** @example 123456 */
            identity_document_number: string;
            identity_document_issue_date: components['schemas']['Date'];
            identity_document_expiry_date: components['schemas']['Date'];
            /** @example [
             *       "Investors",
             *       "Revenue"
             *     ] */
            sources_of_fund: string[];
            addresses: components['schemas']['UpdateEntityAddress'][];
        };
// ...
    };
// ...
}